requirements.txt
# Core dependencies
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1

# HTTP client for Docker Registry API
httpx==0.28.1
aiofiles==24.1.0

# Data validation
pydantic==2.10.5
pydantic-settings==2.7.0

# CORS support
python-multipart==0.0.19

# Testing dependencies (optional, for development)
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0

# Development tools (optional)
black==24.10.0
flake8==7.1.1
mypy==1.14.0